Editor: Use title placeholder in tooltip when post title is empty
When a user opens the post scheduler, we show an indicator on dates with
a currently scheduled post. When the user hovers over those dates, they
see a tooltip with the post title. Currently, that tooltip is empty if
the post has no title. This adds a (No Title) placeholder.

Fixes: #17665.